+++ volmdlr/edges.py
@@ -4904,9 +4904,10 @@ class BSplineCurve3D(BSplineCurve):
         :type parameter: float
         """
         # Is a value of parameter below 4e-3 a real need for precision ?
-        if math.isclose(parameter, 0, abs_tol=4e-3):
+        a, b = self.domain
+        if math.isclose(parameter, a, abs_tol=4e-3):
             return self
-        if math.isclose(parameter, 1, abs_tol=4e-3):
+        if math.isclose(parameter, b, abs_tol=4e-3):
             return self.reverse()
         #     raise ValueError('Nothing will be left from the BSplineCurve3D')
 
@@ -4921,12 +4922,10 @@ class BSplineCurve3D(BSplineCurve):
         :type parameter: float
         """
         # Is a value of parameter below 4e-3 a real need for precision ?
-        if math.isclose(parameter, 0, abs_tol=1e-6):
-            #     # raise ValueError('Nothing will be left from the BSplineCurve3D')
-            #     curves = operations.split_curve(operations.refine_knotvector(self.curve, [4]), parameter)
-            #     return self.from_geomdl_curve(curves[0])
+        a, b = self.domain
+        if math.isclose(parameter, a, abs_tol=1e-6):
             return self.reverse()
-        if math.isclose(parameter, 1, abs_tol=4e-3):
+        if math.isclose(parameter, b, abs_tol=4e-3):
             return self
         curves = volmdlr.nurbs.operations.split_curve(self, round(parameter, 7))
         return curves[0]
@@ -5090,15 +5089,15 @@ class BSplineCurve3D(BSplineCurve):
 
         polys = [volmdlr.wires.ClosedPolygon3D(c.discretization_points(number_points=36)) for c in contours]
 
-        size_v, size_u = len(polys[0].points) + 1, len(polys)
+        size_v, size_u = len(polys[0].points), len(polys)
         degree_u, degree_v = 3, 3
 
         points_3d = []
         for poly in polys:
             points_3d.extend(poly.points)
-            points_3d.append(poly.points[0])
 
-        bezier_surface3d = volmdlr.surfaces.BezierSurface3D(degree_u, degree_v, points_3d, size_u, size_v)
+        bspline_surface3d = volmdlr.surfaces.BSplineSurface3D.from_points_interpolation(points_3d, size_u,
+                                                                                       size_v,degree_u, degree_v)
 
         outer_contour = volmdlr.wires.Contour2D([volmdlr.edges.LineSegment2D(volmdlr.O2D, volmdlr.X2D),
                                                  volmdlr.edges.LineSegment2D(
@@ -5108,7 +5107,7 @@ class BSplineCurve3D(BSplineCurve):
                                                  volmdlr.edges.LineSegment2D(volmdlr.Y2D, volmdlr.O2D)])
         surf2d = volmdlr.surfaces.Surface2D(outer_contour, [])
 
-        bsface3d = volmdlr.faces.BSplineFace3D(bezier_surface3d, surf2d)
+        bsface3d = volmdlr.faces.BSplineFace3D(bspline_surface3d, surf2d)
         new_faces.append(bsface3d)
         return new_faces